fix(app): arrange the code after refactor
