+++ swapi-json-server/package.json
@@ -21,5 +21,8 @@
     "json-server": "0.14.2",
     "lowdb": "1.0.0",
     "serve-static": "1.13.2"
+  },
+  "devDependencies": {
+    "@babel/core": "^7.12.10"
   }
 }
